change project structure
move run.py into /src/lib folder
rename run.py into app.py and added api.py where routers are located
"splited" /predict router into HTMLResponse and JSONResponse type
