+++ Dockerfile
@@ -13,4 +13,4 @@ RUN python -m pip install --no-cache-dir -r requirements.txt -r requirements-dev
 COPY . /app
 
 EXPOSE 8080
-CMD ["python", "run.py"]
+CMD ["fastapi", "run", "./src/lib/app.py"]
