+++ src/lib/template/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{{ title }}</title>
     <link href="{{ url_for('static', path='/styles.css') }}" rel="stylesheet">
-    <link id="favicon" rel="icon" type="image/x-icon" href="static/favicon.ico">
+    <link id="favicon" rel="icon" type="image/x-icon" href="{{ url_for('static', path='/favicon.ico') }}">
 </head>
 <body>
     <div class="container">
